Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    int r;
+    float r;
     scanf("%d",&r);
     printf("Area: %f",r*r*3.14);
     return 0;
